+++ RMZDrv/BinaryReader.h
@@ -2,14 +2,14 @@
 
 typedef struct _BINARYREADER
 {
-	UINT32 currentPosition;
+	UINT64 currentPosition;
 	PUINT8 buffer;
-	UINT32 bufferSize;
+	UINT64 bufferSize;
 } BINARYREADER, *PBINARYREADER;
 
-void RmzBrInit(PBINARYREADER br, PVOID buffer, UINT32 size);
+void RmzBrInit(PBINARYREADER br, PVOID buffer, UINT64 size);
 UINT64 RmzBrReadUInt64(PBINARYREADER br);
 UINT32 RmzBrReadUInt32(PBINARYREADER br);
 UINT16 RmzBrReadUInt16(PBINARYREADER br);
 UINT8 RmzBrReadUInt8(PBINARYREADER br);
-PVOID RmzBrReadBuffer(PBINARYREADER br, UINT32 length, BOOL copy);
+PVOID RmzBrReadBuffer(PBINARYREADER br, UINT64 length, BOOL copy);
